Add Physician entity, controller, repo, service

+++ backend/emergency-room-transfer/src/main/java/group9/emergencyroomtransfer/model/patient/PatientRepository.java
@@ -7,5 +7,5 @@ import java.util.Optional;
 
 @Repository
 public interface PatientRepository extends JpaRepository<Patient, Long> {
-    Optional<Patient> findFirstByPhn(int phn);
+    Optional<Patient> findFirstByPhn(long phn);
 }
